Add .NET8 version of ConditionSet that replaces Expression with DescriptionAttribute for debug metadata
It turns out that compiling expressions at runtime is extremely slow at scale.
Since the only reason expressions were used was to provide debug metadata, we can replace that metadata with a user-defined string. However, because this is info is really only useful when debugging, I didn't want the strings to be part of the class. The built-in DescriptionAttribute provides a clear way to designate the string as metadata, BUT until C#10 you couldn't define attributes on lambda functions, which I expect to be a typical use case for conditions. So I am limiting the Attribute version of ConditionSet to .NET8+, and keeping the old version for previous versions of .NET.